+++ RELEASE_NOTES.md
@@ -3,6 +3,8 @@
 Changes:
 
   * Changed scientific names into a controlled options list
+  * Fixed subproject options not showing up for some projects on bulk Create/Edit Samples page
+  * Fixed subproject validation - if a project has subprojects, its samples must be assigned subprojects
 
 # 1.3.0